Move `Timestamp` to sailfish-types.
