+++ sailfish-types/Cargo.toml
@@ -6,6 +6,7 @@ edition.workspace = true
 rust-version.workspace = true
 
 [dependencies]
+arbitrary = { workspace = true, optional = true }
 arrayvec = { workspace = true }
 async-trait = { workspace = true }
 bincode = { workspace = true }
