+++ timeboost-types/Cargo.toml
@@ -5,6 +5,9 @@ version.workspace = true
 edition.workspace = true
 rust-version.workspace = true
 
+[features]
+arbitrary = ["dep:arbitrary", "sailfish-types/arbitrary"]
+
 [dependencies]
 alloy-consensus = { workspace = true }
 alloy-primitives = { workspace = true }
